\name{agefat}
\alias{agefat}
\docType{data}
\title{Age and body fat percentage}
\description{
In order to relate the body fat percentage to age, researchers selected nine
healthy adults and determined their body fat percentage.
}
\usage{data(agefat)}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{age}}{age of the subject}
    \item{\code{fatpct}}{body fat percentage}
  }
}
%\details{
%
%}
\source{
Ib Skovgaard (2004).Basal Biostatistik 2, Samfundslitteratur.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(agefat)
}
\keyword{datasets}
