\name{aids}
\alias{aids}
\docType{data}
\title{Aids prevalence data}
\description{
Number of aids cases and deaths for a 19-year period.
}
\usage{data(aids)}
\format{
  A data frame with 19 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{a numeric vector}
    \item{\code{cases}}{a numeric vector}
    \item{\code{deaths}}{a numeric vector}
  }
}
%\details{
%
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(aids)
}
\keyword{datasets}
