\name{butterfat}
\alias{butterfat}
\docType{data}
\title{Butterfat and dairy cattle}
\description{
Average butterfat content (percentages)
for random samples of 20 cows (10 two year olds and 10 mature (greater
than four years old)) from each of five breeds.
}
\usage{data(butterfat)}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{Butterfat}}{a numeric vector}
    \item{\code{Breed}}{a factor with levels \code{Ayrshire} \code{Canadian} \code{Guernsey} \code{Holstein-Fresian} \code{Jersey}}
    \item{\code{Age}}{a factor with levels \code{2year} \code{Mature}}
  }
}
%\details{
%
%}
\source{
Hand et al. (1993). A Handbook of Small Data Sets. Chapman and Hall
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(butterfat)
}
\keyword{datasets}
