\name{cabbage}
\alias{cabbage}
\docType{data}
\title{Cabbage yield}
\description{
Cabbage yield for different treatment methods and different fields
}
\usage{data(cabbage)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{method}}{a factor with levels \code{A} \code{C} \code{K} \code{N}}
    \item{\code{yield}}{a numeric vector}
    \item{\code{field}}{a numeric vector}
  }
}
%\details{
%
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(cabbage)
}
\keyword{datasets}
