\name{chicken}
\alias{chicken}
\docType{data}
\title{Weight gain for chickens}
\description{
Twenty chickens were fed with four different feed types - five
chickens for each type - and the weight gain was registered for each
chicken after a period.
}
\usage{data(chicken)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{feed}}{id of feed type. Numeric but it should be used a factor}
    \item{\code{gain}}{Weight gain (numeric)}
  }
}
%\details{
%
%}
\source{
Anonymous (1949). Query 70. Biometrics, 250--251.}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(chicken)
}
\keyword{datasets}
