\name{cornyield}
\alias{cornyield}
\docType{data}
\title{Yield of corn after fertilizer treatment}
\description{
Two varieties of corn were randomly assigned to the 8 plots in a completely
randomized design so that each variety was planted on 4 plots. Four
amounts of fertilizer (5, 10, 15, and 20 units) were randomly 
assigned to the 4 plots in which variety A was planted.  Likewise,
the same four amounts of fertilizer were randomly assigned to the 4
plots in which variety B was planted. Yield in bushels per acre was
recorded for each plot at the end of the experiment.
}
\usage{data(cornyield)}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{\code{yield}}{a numeric vector}
    \item{\code{variety}}{a factor with levels \code{A} \code{B}}
    \item{\code{fertilizer}}{a numeric vector}
  }
}
%\details{
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(cornyield)
}
\keyword{datasets}
