\name{gestation}
\alias{gestation}
\docType{data}
\title{Gestation period for 13 horses}
\description{
The length of the gestation period (the period from conception to birth) was
registered for 13 horses.
}
\usage{data(gestation)}
\format{
  A data frame with 13 observations on the following variable.
  \describe{
    \item{\code{gest}}{length of gestation period}
  }
}
%\details{
%}
\source{
Fictious (but realistic) data.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(gestation)
}
\keyword{datasets}
