\name{hormone}
\alias{hormone}
\docType{data}
\title{Hormone concentration in cattle}
\description{
As part of a larger cattle study, the effect of a particular type of feed on
the concentration of a certain hormone was investigated.  Nine cows were given
the feed for a period, and the hormone concentration was measured initially
and at the end of the period. 
}
\usage{data(hormone)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{feed}}{a numeric vector}
    \item{\code{initial}}{a numeric vector}
    \item{\code{final}}{a numeric vector}
  }
}
%\details{
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(hormone)
}
\keyword{datasets}
