\name{isdals-package}
\alias{isdals-package}
\alias{isdals}
\docType{package}
\title{Data sets for Introduction to Statistical Data Analysis for the Life
Sciences}
\description{
  The package provides the data sets from the book "Introduction to
  Statistical Data Analysis for the Life Sciences" and is also used for
  the course in Statistical Data Analysis 1 held at the University of
  Copenhagen. 
}
\details{
\tabular{ll}{
Package: \tab isdals\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-08-03\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maintainer: Claus Ekstrom <ekstrom@life.ku.dk>
}
\references{
Claus Thorn Ekstrom and Helle Sorensen (2010). Introduction to
Statistical Data Analysis for the Life Sciences. CRC Press 
}
\keyword{ package }
\examples{
data(chicken)
data(tartar)
}
