\name{malaria}
\alias{malaria}
\docType{data}
\title{Parasite counts for children with malaria}
\description{
A medical researcher took blood samples from 31 children who were infected with
malaria and determined for each child the number of malaria parasites in 1 ml
of blood.
}
\usage{data(malaria)}
\format{
  A data frame with 31 observations on the following variable.
  \describe{
    \item{\code{parasites}}{the number of malaria parasites}
  }
}
%\details{
%}
\source{
M.L. Samuels and J.A. Witmer (2003).
Statistics for the Life Sciences (3rd ed.).
Pearson Education, Inc., New Jersey.
}

\references{
C. B. Williams (1964)
Patterns in the Balance of Nature.
Academic Press, London.
}
\examples{
data(malaria)
}
\keyword{datasets}
