\name{mincedmeat}
\alias{mincedmeat}
\docType{data}
\title{Weight of packs with minced meat}
\description{
In meat production, packs of minced meat are specified to contain 500
grams of minced meat.  A sample of ten packs was drawn at random
and the weights (in grams) of the content was recorded.
}
\usage{data(mincedmeat)}
\format{
  A data frame with 10 observations on the following variable.
  \describe{
    \item{\code{wgt}}{weight of minced meat in grams}
  }
}
%\details{
%}
\source{
Fictitious data.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(mincedmeat)
}
\keyword{datasets}
