\name{oilvit}
\alias{oilvit}
\docType{data}
\title{Utilization of vitamin A}
\description{
In an experiment on the utilization of vitamin A, 20 rats were given
vitamin A over a period of three days. Ten rats were
fed vitamin A in corn oil and ten rats were fed vitamin A in
castor oil (American oil).  On the fourth day, the liver of each rat was examined and
the vitamin A concentration in the liver was determined.
}
\usage{data(oilvit)}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{type}}{type of oil. A factor with levels \code{am} \code{corn}}
    \item{\code{avit}}{vitamin A concentration in liver}
  }
}
%\details{
%}
\source{
C.I.Bliss (1967). Statistics in Biology. McGraw-Hill, New York
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(oilvit)
}
\keyword{datasets}
