\name{phosphor}
\alias{phosphor}
\docType{data}
\title{Phosphor concentration in plants during growth}
\description{
In a plant physiological experiment the amount of water-soluble
phosphorous (among others) was measured in the plants, as a percentage
of dry matter.
The phosphorous concentration was measured nine weeks during the growth
season, and the averages over the plants in the experiments was reported.
}
\usage{data(phosphor)}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{week}}{week number}
    \item{\code{phos}}{phosphor concentration (average over the plants)}
  }
}
%\details{
%}
\source{
Ib Skovgaard (2004). Basal Biostatistik 2, Samfundslitteratur.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(phosphor)
}
\keyword{datasets}
