\name{pine}
\alias{pine}
\docType{data}
\title{Height and diameter of pines}
\description{
The data consist of height and diameter (in breast height) measurements from 18 pine trees.
}
\usage{data(pine)}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{diam}}{diameter of the pine tree}
    \item{\code{height}}{height of the pine tree}
  }
}
%\details{
%}
\source{
J.N.R. Jeffers (1959). Experimental Design and Analysis in Forest Research.
Almqvist and Wiksell, Stockholm.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(pine)
}
\keyword{datasets}
