\name{pork}
\alias{pork}
\docType{data}
\title{Pork colour over time}
\description{Investigation of meat quality of pork through
  color stability of pork chops. The color was measured from a pork
  chop from each of ten pigs at days 1, 4, and 6 after storage.
}
\usage{data(pork)}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{brightness}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{pig}}{a numeric vector}
  }
}
%\details{
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(pork)
}
\keyword{datasets}
