\name{ratweight}
\alias{ratweight}
\docType{data}
\title{Weight gain of rats}
\description{
Data contains the weight gain for rats fed on four
different diets: combinations of protein source (beef or cereal) and
protein amount (low and high)
}
\usage{data(ratweight)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{Gain}}{a numeric vector}
    \item{\code{Protein}}{a factor with levels \code{Beef} \code{Cereal}}
    \item{\code{Amount}}{a factor with levels \code{High} \code{Low}}
  }
}
%\details{
%
%}
\source{
  Hand et al. (1993). A Handbook of Small Data Sets. Chapman and Hall
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(ratweight)
}
\keyword{datasets}
