\name{soybean}
\alias{soybean}
\docType{data}
\title{Stress and growth for soybeans}
\description{
An experiment was carried out with 26 soybean plants.  The plants were
pairwise genetically identical, so there were 13 pairs in total. For
each pair, one of the plants was 'stressed' by being shaken daily,
whereas the other plant was not shaken.  After a period the plants
were harvested and the total leaf area was measured for each plant.
}
\usage{data(soybean)}
\format{
  A data frame with 13 observations on the following 3 variables.
  \describe{
    \item{\code{pair}}{id of the pair of plants}
    \item{\code{stress}}{Total leaf area of stressed plant}
    \item{\code{nostress}}{total leaf area of control plant}
  }
}
%\details{
%
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(soybean)
}
\keyword{datasets}
