\name{tartar}
\alias{tartar}
\docType{data}
\title{Tartar for dogs}
\description{
A dog experiment was carried out in order to examine the effect of two
treatments on the development of tartar.
Apart from the two treatment groups there was also a control group.  Twenty-six dogs
were used and allocated to one of the three groups.
After four weeks each dog was examined,
and the development of tartar was summarized by an index.
}
\usage{data(tartar)}
\format{
  A data frame with 26 observations on the following 2 variables.
  \describe{
    \item{\code{treat}}{treatment. A factor with levels \code{Control} \code{HMP} \code{P2O7}}
    \item{\code{index}}{a numeric vector with the tartar index}
  }
}
%\details{
%
%}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(tartar)
}
\keyword{datasets}
