\name{turtles}
\alias{turtles}
\docType{data}
\title{Clutch size of turtles}
\description{
Data to examine the effect of turtle carapace length on
the clutch size of turtles.
}
\usage{data(turtles)}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{length}}{a numeric vector}
    \item{\code{clutch}}{a numeric vector}
  }
}
%\details{
%
%}
\source{
K. G. Ashton and R. L. Burke and J. N. Layne (2007). Geographic
variation in body and clutch size of gopher tortoises. Copeia. 49:355-363.
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(turtles)
## maybe str(turtles) ; plot(turtles) ...
}
\keyword{datasets}
