% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Compute the logit}
\usage{
logit(p)
}
\arguments{
\item{p}{a probability between 0 and 1}
}
\value{
A number with list with class \code{htest} containing the following components:
}
\description{
Compute the logit of a probability
}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk}
}
\keyword{utils}
